% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_releases}
\alias{gh_releases}
\title{List GitHub releases}
\usage{
gh_releases(repo, owner, .limit = Inf)
}
\arguments{
\item{repo}{the name of the repository.}

\item{owner}{the account owner of the repository.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}
}
\value{
data.frame containing releases from the repository.
}
\description{
Get a list of GitHub releases using \code{\link[gh:gh]{gh::gh()}}.
}
\details{
Returns \code{gh_releases()$time} in UTC, set to \code{created_at} for drafts, and \code{published_at} for all
others.
}
