% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\name{render_timevis}
\alias{render_timevis}
\title{Render timeline visualization}
\usage{
render_timevis(data, groups, file, showZoom = FALSE)
}
\arguments{
\item{data}{A dataframe containing the timeline items. Each item on the
timeline is represented by a row in the dataframe. \code{start} and
\code{content} are required for each item, while several other variables
are also supported. See the \strong{Data format} section below for more
details.}

\item{groups}{A dataframe containing the groups data (optional). See the
\strong{Groups} section below for more details.}

\item{file}{File to save HTML into}

\item{showZoom}{If \code{TRUE} (default), then include "Zoom In"/"Zoom Out"
buttons on the widget.}
}
\value{
A timeline visualization \code{htmlwidgets} object
}
\description{
Create a timeline visualization using \code{\link[timevis:timevis]{timevis::timevis()}} and save using
\code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}.
}
\details{
\code{render_timevis()} saves the widget with a fixed width (1400, the width of my laptop screen) and
returns an \code{htmlwidgets} object without a specified width.
}
