% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{siracon2022-package}
\alias{siracon2022}
\alias{siracon2022-package}
\title{siracon2022: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/siracon2022}
  \item \url{https://github.com/jabenninghoff/siracon2022}
  \item Report bugs at \url{https://github.com/jabenninghoff/siracon2022/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
